+++ CHANGELOG.md
@@ -13,6 +13,11 @@
 
 ## vNext
 
+### ✨ New
+
+-   Add `list` command to display list of installed plugins.
+-   Add `unlink` command to unlink installed plugins.
+
 ### 🐞 Fix
 
 -   Resolve DEP0190 warning when using Node.js 24 or higher.